push

+++ 010_randomize.py
@@ -127,15 +127,15 @@ def randomize_two_sample(df):
     rois = ['SUBCORTICAL']
     # Run Randomize
     for roi in rois:
-        if not os.path.isfile('randomise_%s_tstat1.nii.gz'%roi):
+        if not os.path.isfile('randomise_CP_%s_tstat1.nii.gz'%roi):
             print '######################################'
             print 'Running randomise for roi:', roi
             qsm_list = [os.path.join(workspace_iron, subject, 'QSM/QSMnorm_MNI1mm_%s.nii.gz' % roi) for subject in population]
             #print qsm_list
             stats_dir = mkdir_path(os.path.join(ahba_dir,  'RANDOMISE'))
             os.chdir(stats_dir)
-            os.system('fslmerge -t concat_%s.nii.gz %s' % (roi, ' '.join(qsm_list)))
-            os.system('randomise -i concat_%s -o randomise_%s -d design_twosample.mat -t design_twosample.con -R'% (roi, roi))
+            os.system('fslmerge -t concat_CP_%s.nii.gz %s' % (roi, ' '.join(qsm_list)))
+            os.system('randomise -i concat_CP_%s -o randomise_CP_%s -d design_twosample.mat -t design_twosample.con -R'% (roi, roi))
             os.system('rm -rf *concat*')
 
 
@@ -179,18 +179,20 @@ def randomize_one_sample(df):
                       % (control,age, sex, efc, qi1))
         mat.close()
 
+    rois = ['SUBCORTICAL']
     # Run Randomize
     for roi in rois:
-        print '######################################'
-        print 'Running randomise for roi:', roi
-        qsm_list = [os.path.join(workspace_iron, subject, 'QSM/QSMnorm_MNI1mm_%s.nii.gz' % roi) for subject in
+        if not os.path.isfile('randomise_LE_%s_tstat1.nii.gz' % roi):
+            print '######################################'
+            print 'Running randomise for roi:', roi
+            qsm_list = [os.path.join(workspace_iron, subject, 'QSM/QSMnorm_MNI1mm_%s.nii.gz' % roi) for subject in
                     population]
-        # print qsm_list
-        stats_dir = mkdir_path(os.path.join(ahba_dir, 'RANDOMISE'))
-        os.chdir(stats_dir)
-        os.system('fslmerge -t concat_%s.nii.gz %s' % (roi, ' '.join(qsm_list)))
-        os.system('randomise -i concat_%s -o randomise_%s -d design_onesample.mat -t design_onesample.con -R' % (roi, roi))
-        os.system('rm -rf *concat*')
+            # print qsm_list
+            stats_dir = mkdir_path(os.path.join(ahba_dir, 'RANDOMISE'))
+            os.chdir(stats_dir)
+            os.system('fslmerge -t concat_LE_%s.nii.gz %s' % (roi, ' '.join(qsm_list)))
+            os.system('randomise -i concat_LE_%s -o randomise_LE_%s -d design_onesample.mat -t design_onesample.con -R' % (roi, roi))
+            os.system('rm -rf *concat*')
 
 
 ######################################################
